restructured entities and DAO
