+++ flex/src/main/java/consonants/flex/FlexApplication.java
@@ -1,8 +1,5 @@
-package consonants.flex.app;
+package consonants.flex;
 
-import consonants.flex.data_access.mongo_data_access.ClientRepository;
-import consonants.flex.entity.Client;
-import consonants.flex.use_case.create_new_claim.CreateNewClaimDataAccessInterface;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
