+++ flex/src/main/java/consonants/flex/data_access/mongo_data_access/ClaimRepository.java
@@ -1,8 +1,13 @@
 package consonants.flex.data_access.mongo_data_access;
 
 import consonants.flex.entity.Claim;
+import org.bson.types.ObjectId;
 import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
 
-public interface ClaimRepository extends MongoRepository<Claim, Integer> {
+import java.util.List;
+
+@Repository
+public interface ClaimRepository extends MongoRepository<Claim, ObjectId> {
 
 }
