+++ flex/src/main/java/consonants/flex/data_access/mongo_data_access/ClientRepository.java
@@ -1,10 +1,14 @@
 package consonants.flex.data_access.mongo_data_access;
 
 import consonants.flex.entity.Client;
+import org.bson.types.ObjectId;
 import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
-public interface ClientRepository extends MongoRepository<Client, Integer> {
+@Repository
+public interface ClientRepository extends MongoRepository<Client, ObjectId> {
     Optional<Client> findClientById(int clientId);
+
 }
