+++ flex/src/main/java/consonants/flex/data_access/mongo_data_access/FormRepository.java
@@ -1,7 +1,10 @@
 package consonants.flex.data_access.mongo_data_access;
 
 import consonants.flex.entity.Form;
+import org.bson.types.ObjectId;
 import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
 
-public interface FormRepository extends MongoRepository<Form, Integer> {
+@Repository
+public interface FormRepository extends MongoRepository<Form, ObjectId> {
 }
