+++ flex/src/main/java/consonants/flex/entity/Form.java
@@ -2,6 +2,8 @@ package consonants.flex.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import org.bson.types.ObjectId;
+import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 @Document(collection = "forms")
@@ -9,6 +11,8 @@ import org.springframework.data.mongodb.core.mapping.Document;
 @AllArgsConstructor
 public class Form {
 
+    @Id
+    private ObjectId id;
     private boolean confirmed;
     private boolean filledOut;
     private boolean inProgress;
